stabilized implementation of now update pattern
passive membrane diffusion should be working
other modules may need some more work

+++ singa-simulation/src/main/java/de/bioforscher/singa/simulation/features/permeability/MembraneEntry.java
@@ -36,9 +36,9 @@ public class MembraneEntry extends AbstractFeature<Quantity<Frequency>> implemen
         Quantity<Frequency> scaledQuantity = getFeatureContent()
                 .to(new ProductUnit<>(ONE.divide(time.getUnit())));
         // transform to specified amount
-        this.scaledQuantity = scaledQuantity.multiply(time.getValue());
+        this.scaledQuantity = scaledQuantity.multiply(time.getValue().doubleValue());
         // and half
-        this.halfScaledQuantity = scaledQuantity.multiply(time.multiply(0.5).getValue());
+        this.halfScaledQuantity = scaledQuantity.multiply(time.multiply(0.5).getValue().doubleValue());
     }
 
     @Override
